validate chains
