+++ packages/arb-token-bridge-ui/package.json
@@ -52,7 +52,7 @@
   },
   "scripts": {
     "dev": "next dev",
-    "prebuild": "yarn generateDenylist",
+    "prebuild": "yarn validateOrbitChains && yarn generateDenylist",
     "postinstall": "patch-package",
     "build": "next build",
     "start": "next start",
@@ -62,6 +62,7 @@
     "lint:fix": "tsc && eslint 'src/**/*.{js,ts,tsx}' --quiet --fix",
     "prettier:format": "prettier --config-precedence file-override --write \"src/**/*.{tsx,ts,scss,md,json}\"",
     "generateDenylist": "ts-node --project ./scripts/tsconfig.json ./scripts/generateDenylist.ts",
+    "validateOrbitChains": "ts-node --project ./scripts/tsconfig.json ./scripts/validateOrbitChains.ts",
     "e2e:open": "env-cmd -f .e2e.env yarn synpress open --configFile synpress.config.ts",
     "e2e:run": "env-cmd -f .e2e.env yarn synpress run --configFile synpress.config.ts",
     "e2e:run:cctp": "E2E_CCTP=true yarn e2e:run",
